# api/openapi.yaml
openapi: 3.0.3
info:
  title: ComicBagi
  description: PHP Symfony-based comic hosting catalog full-stack.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api
paths:
  /rest/comics:
    get:
      tags:
        - Comic
      summary: List comic.
      operationId: listComic
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: destinationLink
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Comic list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comic'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Comic
      summary: Add comic.
      operationId: addComic
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComic'
        required: true
      responses:
        '201':
          description: Comic added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comic'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{code}:
    get:
      tags:
        - Comic
      summary: Get comic.
      operationId: getComic
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comic gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comic'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Comic
      summary: Update comic.
      operationId: updateComic
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetComic'
        required: true
      responses:
        '200':
          description: Comic updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comic'
        '204':
          description: Comic unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Comic
      summary: Delete comic.
      operationId: deleteComic
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comic deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/destination-links:
    get:
      tags:
        - Comic
      summary: List comic destination link.
      operationId: listComicDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkWebsiteHost
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkRelativeReference
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkHREF
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Comic destination link list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComicDestinationLink'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Comic
      summary: Add comic destination link.
      operationId: addComicDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComicDestinationLink'
        required: true
      responses:
        '201':
          description: Comic destination link added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicDestinationLink'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/destination-links/{ulid}:
    get:
      tags:
        - Comic
      summary: Get comic destination link.
      operationId: getComicDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comic destination link gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicDestinationLink'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Comic
      summary: Update comic destination link.
      operationId: updateComicDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetComicDestinationLink'
        required: true
      responses:
        '200':
          description: Comic destination link updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicDestinationLink'
        '204':
          description: Comic destination link unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Comic
      summary: Delete comic destination link.
      operationId: deleteComicDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comic destination link deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/chapters:
    get:
      tags:
        - Comic Chapter
      summary: List comic chapter.
      operationId: listComicChapter
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: destinationLink
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Comic chapter list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComicChapter'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Comic Chapter
      summary: Add comic chapter.
      operationId: addComicChapter
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComicChapter'
        required: true
      responses:
        '201':
          description: Comic chapter added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapter'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/chapters/{nv}:
    get:
      tags:
        - Comic Chapter
      summary: Get comic chapter.
      operationId: getComicChapter
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: nv
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comic chapter gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapter'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Comic Chapter
      summary: Update comic chapter.
      operationId: updateComicChapter
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: nv
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetComicChapter'
        required: true
      responses:
        '200':
          description: Comic chapter updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapter'
        '204':
          description: Comic chapter unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Comic Chapter
      summary: Delete comic chapter.
      operationId: deleteComicChapter
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: nv
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comic chapter deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/chapters/{chapterNV}/destination-links:
    get:
      tags:
        - Comic Chapter
      summary: List comic chapter destination link.
      operationId: listComicChapterDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: chapterNV
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkWebsiteHost
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkRelativeReference
          in: query
          schema:
            type: array
            items:
              type: string
        - name: linkHREF
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Comic destination link list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComicChapterDestinationLink'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Comic Chapter
      summary: Add comic chapter destination link.
      operationId: addComicChapterDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: chapterNV
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComicChapterDestinationLink'
        required: true
      responses:
        '201':
          description: Comic chapter destination link added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapterDestinationLink'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/comics/{comicCode}/chapters/{chapterNV}/destination-links/{ulid}:
    get:
      tags:
        - Comic Chapter
      summary: Get comic chapter destination link.
      operationId: getComicChapterDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: chapterNV
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comic chapter destination link gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapterDestinationLink'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Comic Chapter
      summary: Update comic chapter destination link.
      operationId: updateComicChapterDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: chapterNV
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetComicChapterDestinationLink'
        required: true
      responses:
        '200':
          description: Comic chapter destination link updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapterDestinationLink'
        '204':
          description: Comic chapter destination link unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Comic Chapter
      summary: Delete comic chapter destination link.
      operationId: deleteComicChapterDestinationLink
      parameters:
        - name: comicCode
          in: path
          required: true
          schema:
            type: string
        - name: chapterNV
          in: path
          description: Number[+Version]
          required: true
          schema:
            type: string
        - name: ulid
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comic chapter destination link deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
  /rest/links:
    get:
      tags:
        - Link
      summary: List link.
      operationId: listLink
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: websiteHost
          in: query
          schema:
            type: array
            items:
              type: string
        - name: relativeReference
          in: query
          schema:
            type: array
            items:
              type: string
        - name: href
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Link list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Link
      summary: Add link.
      operationId: addLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLink'
        required: true
      responses:
        '201':
          description: Link added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/links/{href}:
    get:
      tags:
        - Link
      summary: Get link.
      operationId: getLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Link
      summary: Update link.
      operationId: updateLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLink'
        required: true
      responses:
        '200':
          description: Link updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '204':
          description: Link unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Link
      summary: Delete link.
      operationId: deleteLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Link deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
components:
  schemas:
    Comic:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        code:
          type: string
        destinationLinkCount:
          type: integer
        chapterCount:
          type: integer
      required:
        - createdAt
        - code
        - destinationLinkCount
        - chapterCount
    NewComic:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    SetComic:
      type: object
      properties:
        code:
          type: string
          nullable: true
    ComicDestinationLink:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        ulid:
          type: string
          x-go-name: ULID
        linkWebsiteHost:
          type: string
        linkWebsiteName:
          type: string
        linkRelativeReference:
          type: string
          nullable: true
        releasedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - createdAt
        - ulid
        - linkWebsiteHost
        - linkWebsiteName
    NewComicDestinationLink:
      type: object
      properties:
        linkWebsiteHost:
          type: string
        linkRelativeReference:
          type: string
          nullable: true
        releasedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - linkWebsiteHost
...
    NewComicChapter:
      type: object
      properties:
        number:
          type: number
        version:
          type: string
          nullable: true
      required:
        - number
...
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT